+++ LowLevelSupport~/Unity.ZeroJobs/AssemblyInfo.cs
@@ -6,3 +6,4 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Unity.ZeroJobs.Tests")]
 [assembly: InternalsVisibleTo("Unity.Collections")]
 [assembly: InternalsVisibleTo("Unity.Tiny.Rendering.Native")]
+[assembly: InternalsVisibleTo("Unity.Logging")]