+++ LowLevelSupport~/UnsafeUtilityPatcher/Options.cs
@@ -824,21 +824,6 @@ namespace Mono.Options
 			throw new InvalidOperationException("Option has no names!");
 		}
 
-		[Obsolete("Use KeyedCollection.this[string] instead. (RemovedAfter 2020-04-09)")]
-		protected Option GetOptionForName(string option)
-		{
-			if (option == null)
-				throw new ArgumentNullException("option");
-			try
-			{
-				return base[option];
-			}
-			catch (KeyNotFoundException)
-			{
-				return null;
-			}
-		}
-
 		protected override void InsertItem(int index, Option item)
 		{
 			base.InsertItem(index, item);