+++ Unity.Entities.Runtime.Build/AssemblyInfo.cs
@@ -3,3 +3,4 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Unity.Dots.TestRunner.Editor")]
 [assembly: InternalsVisibleTo("Unity.Entities.Runtime.Build.Tests")]
 [assembly: InternalsVisibleTo("Unity.Entities.Runtime.Build.Editor")]
+[assembly: InternalsVisibleTo("Unity.DevMode.DotsRuntimeSupport")]