+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.26.0] - 2020-05-15
+* Fixes the `OutputBuildDirectory` BuildComponent to accept absolute paths. Previously builds would fail whenever an absolute path was used.
+* Removed expired API `OptionSet.GetOptionForName(string option)`
+* Updated minimum Unity Editor version to 2019.3.12f1 (84b23722532d)
+
 ## [0.25.0] - 2020-04-30
 * Added an error message guiding the user to fix their BuildConfiguration asset if the asset name contains a space, rather than having the build fail with a seemingly unrelated error.
 * Fixed an issue where the first component defined in an assembly would not get a generated Equals() and GetHashCode() implementation which could result in incorrect runtime behaviour as components could not be properly compared.